main.py argument additions; server log name change
Added a possibility to specify the remote server IP and port on main.py.
Now it's easier to set up servers on different machines. Also added a
flag -d for starting as daemon (non-daemon by default now). In the log
file name, server IP and port is now separated by _ for clarity. Added
IP and server listen port to the server startup message.